pyproject.toml
[project]
name = "resume-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "PyPDF2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
